images/adapter-claude/requirements.txt
claude_agent_sdk==0.1.6
pyyaml
